requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
discord==1.7.3
discord.py==1.7.3
idna==3.2
multidict==5.1.0
numpy==1.21.2
pygame==2.0.1
python-dotenv==0.19.0
requests==2.26.0
riotwatcher==3.1.4
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3